+++ backend/config/initializers/cors.rb
@@ -2,7 +2,7 @@
 
 Rails.application.config.middleware.insert_before 0, Rack::Cors do
   allow do
-    origins "https://swi-ng.com"
+    origins "https://swi-ng.com", "http://localhost:8000", "http://localhost:4000"
     resource "*",
              headers: :any,
              methods: [:get, :post, :put, :patch, :delete, :options, :head],
