+++ nginx/dev.conf
@@ -1,0 +1,36 @@
+server {
+    listen 80;
+    server_name localhost;
+
+    location / {
+        proxy_pass http://frontend:8000;
+        proxy_set_header Host $host;
+        proxy_set_header X-Real-IP $remote_addr;
+        proxy_set_header X-Forwarded-For $proxy_add_x_forwarded_for;
+        proxy_set_header X-Forwarded-Proto $scheme;
+        add_header 'Access-Control-Allow-Origin' 'http://localhost:8000';
+        add_header 'Access-Control-Allow-Credentials' 'true';
+        add_header 'Access-Control-Allow-Methods' 'GET, OPTIONS, PATCH, DELETE, POST, PUT';
+        add_header 'Access-Control-Allow-Headers' 'X-CSRF-Token, X-Requested-With, X-HTTP-Method-Override, Content-Type, Accept';
+    }
+
+    location /api {
+        proxy_pass http://backend:4000;
+        proxy_set_header Host $host;
+        proxy_set_header X-Real-IP $remote_addr;
+        proxy_set_header X-Forwarded-For $proxy_add_x_forwarded_for;
+        proxy_set_header X-Forwarded-Proto $scheme;
+        add_header 'Access-Control-Allow-Origin' 'http://localhost:8000';
+        add_header 'Access-Control-Allow-Credentials' 'true';
+        add_header 'Access-Control-Allow-Methods' 'GET, OPTIONS, PATCH, DELETE, POST, PUT';
+        add_header 'Access-Control-Allow-Headers' 'X-CSRF-Token, X-Requested-With, X-HTTP-Method-Override, Content-Type, Accept';
+    }
+
+    location /_next/webpack-hmr {
+        proxy_pass http://frontend:8000/_next/webpack-hmr;
+        proxy_http_version 1.1;
+        proxy_set_header Upgrade $http_upgrade;
+        proxy_set_header Connection "Upgrade";
+        proxy_set_header Host $host;
+    }
+}
